payment/pyproject.toml
[tool.poetry]
name = "payment"
version = "0.1.0"
description = ""
authors = ["Mekni_Wassime <mekni.wassime@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"
pillow = "^9.4.0"
requests = "^2.28.2"
coreapi = "^2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
